backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.3.0
azure-cosmos>=4.7.0
azure-functions>=1.20.0
httpx>=0.27.0
python-dotenv>=1.0.1
azure-search-documents>=11.4.0
openai>=1.0.0


